unifying parsing code with MessageFormatter
