+++ kifi-backend/modules/eliza/app/com/keepit/eliza/commanders/ElizaEmailCommander.scala
@@ -34,11 +34,7 @@ import com.keepit.eliza.mail.DomainToNameMapper
 import scala.util.{Failure, Success}
 import com.keepit.common.logging.Logging
 import scala.util.matching.Regex.Match
-
-abstract class MessageSegment(val kind: String, val txt: String) //for use in templates since you can't match on type (it seems)
-case class TextLookHereSegment(override val txt: String, pageText: String) extends MessageSegment("tlh", txt)
-case class ImageLookHereSegment(override val txt: String, imgUrl: String) extends MessageSegment("ilh", txt)
-case class TextSegment(override val txt: String) extends MessageSegment("txt", txt)
+import com.keepit.eliza.util.{MessageFormatter, TextSegment}
 
 class ElizaEmailCommander @Inject() (
     shoebox: ShoeboxServiceClient,
@@ -139,7 +135,7 @@ class ElizaEmailCommander @Inject() (
     }
 
     relevantMessages.filterNot(_.from.isSystem).map{ message =>
-      val messageSegments = ElizaEmailCommander.parseMessage(message.messageText)
+      val messageSegments = MessageFormatter.parseMessageSegments(message.messageText)
       message.from match {
         case MessageSender.User(id) => ExtendedThreadItem(allUsers(id).shortName, allUsers(id).fullName, Some(allUserImageUrls(id)), messageSegments)
         case MessageSender.NonUser(nup) => {
@@ -278,38 +274,6 @@ class ElizaEmailCommander @Inject() (
 }
 
 object ElizaEmailCommander {
-
-  def parseMessage(msg: String): Seq[MessageSegment] = {
-
-    def unescape(string: String) = string.replace(raw"\]","]").replace(raw"\)",")")
-    def parseSegment(m: Match) = {
-      val segments = m.group(3).split('|')
-      val kind = segments.head
-      val payload = unescape(URLDecoder.decode(segments.last, "UTF-8"))
-      val text = unescape(m.group(1))
-      kind match {
-        case "i" => ImageLookHereSegment(text, payload)
-        case "r" => TextLookHereSegment(text, payload)
-        case _ => throw new Exception("Unknown look-here type: " + kind)
-      }
-    }
-
-    try {
-      val re = """\[((?:(?:[^\]]*)\\\])*(?:[^\]]*)+)\](\(x-kifi-sel:((?:(?:[^\)]*)\\\))*(?:[^\)]*)+)\))""".r
-      val (position, segments) = re.findAllMatchIn(msg).foldLeft((0,Seq[MessageSegment]())){ (acc, m) =>
-        val (currPos, seq) = acc
-        val lookHereSegment = parseSegment(m)
-        (m.end, if (m.start > currPos) seq :+ TextSegment(unescape(msg.substring(currPos, m.start))) :+ lookHereSegment
-                else seq :+ lookHereSegment)
-      }
-      val ending = TextSegment(unescape(msg.substring(position)))
-      if (ending.txt.length > 0) segments :+ ending else segments
-    } catch {
-      case t: Throwable => {
-        throw new Exception(s"Exception during parsing of message $msg. Exception was $t")
-      }
-    }
-  }
 
   /**
    * This function is meant to be used from the console, to see how emails look like without deploying to production
