+++ kifi-backend/modules/eliza/app/views/emailImageLookHereBlock.scala.html
@@ -1,4 +1,4 @@
-@(lookerName: String, segment: com.keepit.eliza.commanders.ImageLookHereSegment)
+@(lookerName: String, segment: com.keepit.eliza.util.ImageLookHereSegment)
   <tr>
     <td><table width="100%" border="0" cellspacing="0" cellpadding="0">
         <tr>
