+++ kifi-backend/modules/eliza/app/views/emailMessageBlock.scala.html
@@ -1,4 +1,4 @@
-@(imageUrl: Option[String], shortName: String, fullName: String, segments: Seq[com.keepit.eliza.commanders.MessageSegment])
+@(imageUrl: Option[String], shortName: String, fullName: String, segments: Seq[com.keepit.eliza.util.MessageSegment])
   <tr>
     <td bgcolor="#f3f3f3" class="aside"><table width="100%" border="0" cellspacing="0" cellpadding="0">
         <tr>
@@ -39,10 +39,10 @@
                                     </tr>
                                   }
                                   @if(segment.kind=="tlh") {
-                                    @emailTextLookHereBlock(shortName, segment.asInstanceOf[com.keepit.eliza.commanders.TextLookHereSegment])
+                                    @emailTextLookHereBlock(shortName, segment.asInstanceOf[com.keepit.eliza.util.TextLookHereSegment])
                                   }
                                   @if(segment.kind=="ilh"){
-                                    @emailImageLookHereBlock(shortName, segment.asInstanceOf[com.keepit.eliza.commanders.ImageLookHereSegment])
+                                    @emailImageLookHereBlock(shortName, segment.asInstanceOf[com.keepit.eliza.util.ImageLookHereSegment])
                                   }
                                   <tr>
                                     <td height="5" style="line-height:1px;font-size:1px;">&nbsp;</td>
