+++ kifi-backend/modules/eliza/app/views/emailTextLookHereBlock.scala.html
@@ -1,4 +1,4 @@
-@(lookerName: String, segment: com.keepit.eliza.commanders.TextLookHereSegment)
+@(lookerName: String, segment: com.keepit.eliza.util.TextLookHereSegment)
   <tr>
     <td><table width="100%" border="0" cellspacing="0" cellpadding="0">
         <tr>
