+++ kifi-backend/modules/eliza/app/views/next/emailMessageBlock.scala.html
@@ -1,4 +1,4 @@
-@(imageUrl: Option[String], shortName: String, fullName: String, segments: Seq[com.keepit.eliza.commanders.MessageSegment])
+@(imageUrl: Option[String], shortName: String, fullName: String, segments: Seq[com.keepit.eliza.util.MessageSegment])
 <tr>
   <td bgcolor="#f3f3f3" class="aside"><table width="100%" border="0" cellspacing="0" cellpadding="0">
       <tr>
